Same set of file in Heroku master
These set of file is required by Heroku and will be first push to Heroku master github so that our web app in hosted by Heroku can read the code from Heroku master. Any changes to this repository will be affect the web app. In order to make changes to the web app, we need to go to 
Web app can be accessed using the link below.
https://klangvalley-house.herokuapp.com/